Allow attribute vs property value
